[SPARK-10087] [CORE] [BRANCH-1.5] Disable spark.shuffle.reduceLocality.enabled by default.
https://issues.apache.org/jira/browse/SPARK-10087

In some cases, when spark.shuffle.reduceLocality.enabled is enabled, we are scheduling all reducers to the same executor (the cluster has plenty of resources). Changing spark.shuffle.reduceLocality.enabled to false resolve the problem.

Comments of https://github.com/apache/spark/pull/8280 provide more details of the symptom of this issue.

This PR changes the default setting of `spark.shuffle.reduceLocality.enabled` to `false` for branch 1.5.

Author: Yin Huai <yhuai@databricks.com>

Closes #8296 from yhuai/setNumPartitionsCorrectly-branch1.5.